pyproject.toml
[tool.poetry]
name = "nats_queue"
version = "1.1.4"
description = ""
authors = ["Kristina Shishkina <k.shishkina@darwinsoft.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
nats-py = "^2.9.0"
pytest-asyncio = "^0.24.0"
mock = "^5.1.0"
python-dotenv = "^1.0.1"
coverage = "^7.6.1"
setuptools = "^75.6.0"
wheel = "^0.45.1"
twine = "^6.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"
tox = "^4.21.0"
pytest-cov = "^5.0.0"
pytest = "^8.3.0" 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
